Move EVM tests
