+++ test-utils/testlib/src/lib.rs
@@ -18,6 +18,8 @@ pub mod actix_utils;
 pub mod fees_utils;
 pub mod node;
 pub mod runtime_utils;
+#[cfg(feature = "protocol_feature_evm")]
+pub mod standard_evm_cases;
 pub mod standard_test_cases;
 pub mod test_helpers;
 pub mod user;
